Reorder character creation steps
